move common functions for eval to eval/utils.py
